Explcitly state that "public" allows responses with Authentication to be stored
In #161, the suggestion was made to explicitly note when a directive
allows the response to be cached, even with Authorization being used.

According to section `caching.authenticated.responses`, `public` also
has this effect, so I added a note there.

Additionally, in `cache-response-directive.s-maxage`, I tried to
make it clearer that `s-maxage` overrides Authorization as well,
due to the implication of proxy-revalidate and thus must-revalidate.