IP address reference identities
This is how it works in practice.

RFC 6125 is due for an update, I guess.  iPAddress subjectAltName has been
supported for years and years, but it never managed to capture it properly.

This is definitely a big change in that it now outlaws CN-ID.  And the
definition of IP-ID here is probably overreach, but I don't see how we can
avoid that.  I don't want to update RFC 6125 here (as much as it needs
updating), so making the definition standalone seemed like the right way to
handle it.

Closes #680.