Remove 6919 language
This isn't a place for equivocation.

This could have been MUST except that that isn't good to use here.

I considered just "conforms to" as well.  That would also work, but that is
incompatible with "wishes to" and I was less confident in removing that.  I
could do "conforms to" with "gateway that interoperates with inbound HTTP
servers", but that's a bigger change.