Try to better define "strong" encryption
In the definition of the https scheme, the phrase "strong encryption" is used.
This implies some sort of judgment.  I think that the key is that both
client and server need to agree on the adequacy of the protections.  Noting
that this might be contextual is somewhat obvious at that point, but I thought
it might be worth pointing out that what is generally acceptable might change
over time.