Make Host directly mandatory
The normative language in -semantics wasn't direct enough about the
requirement for Host to be present in requests.  There was good language in
-messaging that I eventually found though, so there isn't any real problem
here, just an opportunity to improve.

I did this by taking the SHOULD...unless text and making that MUST unless,
splitting off the SHOULD that relates to putting Host as the first field line
of the header section.

Also, add a note in -messaging about supplying a default value for the
authority.

Closes #722.