The extent of a protection space is unknowable
When the text suggests that credentials can be automatically applied to all
requests made within a protection space, that begs the question: how do I know
what requests are in the same protection space?

The answer is "well, you don't really".  Authentication schemes might define
something, but otherwise you are left to guess.  This PR says that as directly
as I could manage.

I considered adding another sentence here that says "In the absence of
specific information about the extent of a protection space, clients &MAY;
assume that the protection space extent is the origin of the server."  I'd
like thoughts on whether that is helpful.

Closes #710.