Try to wordsmith the TLS secure close depiction
TLS requires that you send an alert (whether error or "close_notify")
before closing the write side of your connection, but not to wait around
to receive the one from the peer.

Use wording a bit stronger than "provides a facility for" to try to
align more closely with the protocol requirements.