Avoid bare 'for the sake of security'
The mere act of inserting a proxy into the chain does not, in and of
itself, do much of anything for the security of the system (and a badly
implemented proxy can make the security of the system much worse).
A proxy can, however, provide security services such as auditing
access, annotating content from untrustworthy sources, exfiltration
avoidance, etc.  It is these services that are the security-related
motivation for using a proxy, so say that "security services", rather
than just "security", are being provided by the proxy.