wordsmith automatic-link-editing guidance
The current text of "ought to ... where possible" is easy to read
as saying "if you have the technical capability, do it".
There are, however, some subtleties, such as if you get the 308 response
over http-not-s, in which case the question of whether or not to
re-link is not so clear-cut.

Change from "where possible" to "where appropriate" to hint that there
is some logic needed here beyond "blindly accept".